move ExtendedFeatureVessel to app responsability
