+++ applications/fishing-map/src/utils/info.ts
@@ -1,5 +1,5 @@
 import { Vessel } from '@globalfishingwatch/api-types/dist'
-import { ExtendedFeatureVessel } from '@globalfishingwatch/react-hooks'
+import { ExtendedFeatureVessel } from 'features/map/map.slice'
 import i18n from '../features/i18n/i18n'
 
 export const formatInfoField = (fieldValue: string, type: string) => {
