+++ packages/react-hooks/src/index.ts
@@ -11,7 +11,6 @@ export type {
   InteractionEvent,
   InteractionEventCallback,
   ExtendedFeature,
-  ExtendedFeatureVessel,
 } from './use-map-interaction/index'
 export { default as useMapLegend } from './use-map-legend'
 export { default as useTilesState } from './use-tiles-state'
